requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-Cors==4.0.0

# Data science and ML dependencies (FULL VERSION)
scikit-learn
sentence-transformers
numpy

# Environment and configuration
python-dotenv==1.0.0

# OpenAI API
openai

# File monitoring
watchdog

# Cloud database (Supabase)
supabase==2.16.0

# Additional dependencies for cloud deployment
gunicorn
requests
urllib3

# Additional ML and data processing libraries
torch
transformers
scipy

# Authentication and security
PyJWT
cryptography
bcrypt

# Clerk Authentication (using latest stable version)
clerk-backend-sdk>=1.1.0 